+++ bin/user/tests/test_sql_statements.py
@@ -1,0 +1,264 @@
+#
+#    Copyright (c) 2025 Rich Bell <bellrichm@gmail.com>
+#
+#    See the file LICENSE.txt for your full rights.
+#
+
+# pylint: disable=wrong-import-order
+# pylint: disable=missing-docstring
+# pylint: disable=invalid-name
+
+import unittest
+import mock
+
+import configobj
+import io
+import random
+import string
+
+import weeutil.weeutil
+import weewx.manager
+
+import user.aqitype
+
+#usUnits = 1
+archive_interval = 5
+archive_interval_seconds = archive_interval * 60
+
+expected_pm_values = [1.0137931034482757, 0.9433333333333332, 0.996551724137931, 2.103333333333333, 1.4133333333333329,
+                      1.7999999999999996, 1.9499999999999993, 2.617241379310345, 2.433333333333333, 1.7933333333333323,
+                      1.5758620689655178, 1.9233333333333351, 0.7827586206896553, 1.3566666666666667, 1.3566666666666662,
+                      1.1862068965517243, 1.4333333333333325, 0.8965517241379313, 1.0699999999999998, 1.49,
+                      1.317241379310345, 1.63, 1.5999999999999992, 0.8724137931034481, 0.9233333333333327,
+                      0.9206896551724134, 1.806666666666666, 1.329999999999999, 0.8275862068965519, 1.3433333333333326,
+                      1.8275862068965523, 1.3966666666666663, 1.166666666666667, 1.0862068965517244, 1.3799999999999994,
+                      1.5433333333333328, 1.5344827586206886, 0.9366666666666668, 1.6689655172413793, 2.159999999999999,
+                      1.3103448275862069, 1.353333333333333, 1.6966666666666668, 1.8896551724137935, 1.24,
+                      1.7413793103448285, 1.4133333333333344, 1.11, 1.4517241379310344, 1.3300000000000003,
+                      1.282758620689655, 1.23, 0.9933333333333333, 1.0172413793103448, 1.1799999999999997,
+                      1.2433333333333327, 1.031034482758621, 1.3833333333333333, 1.4655172413793105, 1.2033333333333336,
+                      0.9466666666666665, 1.0482758620689654, 1.2500000000000004, 1.589655172413794, 1.3600000000000003,
+                      1.355172413793103, 1.2333333333333334, 1.4500000000000004, 1.3413793103448275, 0.8300000000000003,
+                      1.151724137931035, 0.8233333333333339, 1.4333333333333327, 1.968965517241379, 1.413333333333333,
+                      1.3551724137931038, 2.0000000000000004, 1.2399999999999998, 1.6068965517241385, 1.2433333333333327,
+                      1.0566666666666664, 1.310344827586207, 1.0413793103448277, 1.1758620689655177, 1.06,
+                      1.4100000000000004, 1.4482758620689649, 1.28, 1.7448275862068967, 1.3100000000000007,
+                      1.7199999999999998, 1.820689655172414, 1.6800000000000002, 1.3068965517241382, 0.9233333333333337,
+                      1.0266666666666668, 2.334482758620689, 1.1266666666666665, 1.5931034482758624, 1.5700000000000007,
+                      1.0862068965517244, 2.046666666666666, 1.8333333333333326, 1.4034482758620692, 1.153333333333333,
+                      1.8862068965517236, 1.6233333333333337, 1.4866666666666672, 1.806896551724137, 1.406666666666667,
+                      1.1000000000000003, 0.9999999999999996, 1.3433333333333335, 1.6137931034482755, 1.0233333333333332,
+                      1.7517241379310349, 0.9266666666666666, 1.7, 1.013793103448276, 1.3433333333333326,
+                      1.8448275862068964, 1.5666666666666675, 1.716666666666667, 1.5344827586206902, 1.0433333333333332,
+                      1.103448275862069, 1.5100000000000005, 1.524137931034483, 1.5100000000000007, 2.1833333333333327,
+                      1.4482758620689653, 1.466666666666666, 1.3766666666666665, 1.8965517241379304, 1.836666666666666,
+                      1.3299999999999998, 0.8724137931034479, 1.3900000000000001, 1.8482758620689657, 1.4733333333333332,
+                      0.8400000000000005, 1.3103448275862064, 1.4833333333333336, 1.3275862068965527, 1.4833333333333338,
+                      1.5620689655172417, 1.5166666666666657, 1.7833333333333334, 2.075862068965517, 1.7966666666666662,
+                      1.276666666666667, 2.213793103448276, 1.6666666666666667, 1.582758620689655, 1.4933333333333334,
+                      2.1379310344827576, 1.6900000000000004, 1.4799999999999995, 1.5034482758620689, 1.1599999999999995,
+                      1.6, 1.7034482758620688, 0.9966666666666667, 1.3620689655172409, 1.2699999999999998,
+                      1.3899999999999997, 2.2586206896551717, 1.77, 2.582758620689655, 2.0566666666666666,
+                      1.1566666666666667, 1.510344827586207, 1.3599999999999999, 0.6699999999999999, 0.9586206896551727,
+                      1.8100000000000007, 1.2666666666666668, 1.4931034482758618, 1.0433333333333332, 1.3793103448275856,
+                      1.8366666666666664, 1.7733333333333343, 1.1137931034482755, 1.1799999999999997, 0.9344827586206894,
+                      1.1699999999999997, 1.1482758620689653, 2.1666666666666656, 1.3400000000000003, 1.2965517241379312,
+                      1.2833333333333337, 1.5699999999999998, 1.406896551724138, 1.3166666666666667, 1.1137931034482758,
+                      0.9166666666666667, 1.4000000000000004, 1.2499999999999993, 1.7310344827586197, 1.423333333333334,
+                      1.1299999999999997, 1.3758620689655177, 2.2166666666666672, 1.2310344827586208, 0.7533333333333332,
+                      1.1, 1.5241379310344825, 0.8566666666666666, 1.5379310344827581, 1.6833333333333325,
+                      1.241379310344827, 0.7266666666666668, 1.7033333333333334, 0.9999999999999997, 1.2766666666666662,
+                      1.6566666666666665, 1.2413793103448276, 2.0866666666666664, 1.603448275862069, 1.5799999999999996,
+                      0.9793103448275862, 1.5233333333333332, 1.4000000000000001, 1.0896551724137926, 1.693333333333334,
+                      1.3724137931034481, 1.0333333333333332, 0.9034482758620687, 1.01, 0.8965517241379305,
+                      1.0433333333333334, 1.4433333333333331, 1.327586206896552, 1.1933333333333342, 1.006896551724138,
+                      0.9066666666666666, 0.7620689655172413, 0.9233333333333332, 1.0379310344827586, 0.6733333333333335,
+                      1.275862068965517, 1.5866666666666667, 1.7333333333333334, 1.706896551724138, 1.333333333333333,
+                      0.9310344827586209, 1.416666666666667, 0.9586206896551721, 1.2266666666666666, 1.280000000000001,
+                      1.386206896551724,1.5666666666666667, 1.6866666666666665, 1.3172413793103444, 1.7551724137931037,
+                      1.1166666666666667, 1.2034482758620688, 0.8400000000000001, 1.1482758620689655, 1.1566666666666667,
+                      1.8866666666666674, 1.3551724137931034, 1.366666666666667, 2.382758620689655, 1.2099999999999993,
+                      1.1199999999999997, 1.3068965517241375, 1.1266666666666667, 0.9275862068965519, 1.3733333333333342,
+                      1.0758620689655174, 1.5099999999999996, 1.1733333333333331, 1.8137931034482762, 1.843333333333334,
+                      1.6655172413793105, 2.070000000000001, 1.6206896551724146, 1.2300000000000009, 1.2266666666666661,
+                      1.3344827586206895, 1.7266666666666663, 1.4137931034482762, 1.7999999999999998, 1.5900000000000005,
+                      1.8689655172413788, 1.9833333333333327, 1.5517241379310336]
+
+expected_timestamps = [1740114300, 1740114600, 1740114900, 1740115200, 1740115500, 1740115800, 1740116100, 1740116400, 1740116700, 1740117000,
+                       1740117300, 1740117600, 1740117900, 1740118200, 1740118500, 1740118800, 1740119100, 1740119400, 1740119700, 1740120000,
+                       1740120300, 1740120600, 1740120900, 1740121200, 1740121500, 1740121800, 1740122100, 1740122400, 1740122700, 1740123000,
+                       1740123300, 1740123600, 1740123900, 1740124200, 1740124500, 1740124800, 1740125100, 1740125400, 1740125700, 1740126000,
+                       1740126300, 1740126600, 1740126900, 1740127200, 1740127500, 1740127800, 1740128100, 1740128400, 1740128700, 1740129000,
+                       1740129300, 1740129600, 1740129900, 1740130200, 1740130500, 1740130800, 1740131100, 1740131400, 1740131700, 1740132000,
+                       1740132300, 1740132600, 1740132900, 1740133200, 1740133500, 1740133800, 1740134100, 1740134400, 1740134700, 1740135000,
+                       1740135300, 1740135600, 1740135900, 1740136200, 1740136500, 1740136800, 1740137100, 1740137400, 1740137700, 1740138000,
+                       1740138300, 1740138600, 1740138900, 1740139200, 1740139500, 1740139800, 1740140100, 1740140400, 1740140700, 1740141000,
+                       1740141300, 1740141600, 1740141900, 1740142200, 1740142500, 1740142800, 1740143100, 1740143400, 1740143700, 1740144000,
+                       1740144300, 1740144600, 1740144900, 1740145200, 1740145500, 1740145800, 1740146100, 1740146400, 1740146700, 1740147000,
+                       1740147300, 1740147600, 1740147900, 1740148200, 1740148500, 1740148800, 1740149100, 1740149400, 1740149700, 1740150000,
+                       1740150300, 1740150600, 1740150900, 1740151200, 1740151500, 1740151800, 1740152100, 1740152400, 1740152700, 1740153000,
+                       1740153300, 1740153600, 1740153900, 1740154200, 1740154500, 1740154800, 1740155100, 1740155400, 1740155700, 1740156000,
+                       1740156300, 1740156600, 1740156900, 1740157200, 1740157500, 1740157800, 1740158100, 1740158400, 1740158700, 1740159000,
+                       1740159300, 1740159600, 1740159900, 1740160200, 1740160500, 1740160800, 1740161100, 1740161400, 1740161700, 1740162000,
+                       1740162300, 1740162600, 1740162900, 1740163200, 1740163500, 1740163800, 1740164100, 1740164400, 1740164700, 1740165000,
+                       1740165300, 1740165600, 1740165900, 1740166200, 1740166500, 1740166800, 1740167100, 1740167400, 1740167700, 1740168000,
+                       1740168300, 1740168600, 1740168900, 1740169200, 1740169500, 1740169800, 1740170100, 1740170400, 1740170700, 1740171000,
+                       1740171300, 1740171600, 1740171900, 1740172200, 1740172500, 1740172800, 1740173100, 1740173400, 1740173700, 1740174000,
+                       1740174300, 1740174600, 1740174900, 1740175200, 1740175500, 1740175800, 1740176100, 1740176400, 1740176700, 1740177000,
+                       1740177300, 1740177600, 1740177900, 1740178200, 1740178500, 1740178800, 1740179100, 1740179400, 1740179700, 1740180000,
+                       1740180300, 1740180600, 1740180900, 1740181200, 1740181500, 1740181800, 1740182100, 1740182400, 1740182700, 1740183000,
+                       1740183300, 1740183600, 1740183900, 1740184200, 1740184500, 1740184800, 1740185100, 1740185400, 1740185700, 1740186000,
+                       1740186300, 1740186600, 1740186900, 1740187200, 1740187500, 1740187800, 1740188100, 1740188400, 1740188700, 1740189000,
+                       1740189300, 1740189600, 1740189900, 1740190200, 1740190500, 1740190800, 1740191100, 1740191400, 1740191700, 1740192000,
+                       1740192300, 1740192600, 1740192900, 1740193200, 1740193500, 1740193800, 1740194100, 1740194400, 1740194700, 1740195000,
+                       1740195300, 1740195600, 1740195900, 1740196200, 1740196500, 1740196800, 1740197100, 1740197400, 1740197700, 1740198000,
+                       1740198300, 1740198600, 1740198900, 1740199200, 1740199500, 1740199800, 1740200100, 1740200400]
+
+MANAGER_DICT = '''
+WEEWX_ROOT = 'bin/user/tests/'
+
+#   This section binds a data store to a database.
+[DataBindings]
+    [[aqi_binding]]
+        database = aqi
+        table_name = archive
+        manager = weewx.wxmanager.WXDaySummaryManager
+        schema = schemas.wview_extended.schema
+    
+#   This section defines various databases.
+[Databases]
+    [[aqi]]
+        database_name = pm.sdb
+        database_type = SQLite
+
+#   This section defines defaults for the different types of databases.
+[DatabaseTypes]
+    [[SQLite]]
+        driver = weedb.sqlite
+        SQLITE_ROOT = data
+        '''
+
+def random_string(length=32):
+    return ''.join([random.choice(string.ascii_letters + string.digits) for n in range(length)]) # pylint: disable=unused-variable
+
+# ToDo: is there a better way?
+def mock_calculate_effect(_1, _2, _3, _4):
+    return random.randint(1, 100)
+
+def setup_config(calculated_field, input_field, algorithm, aqi_type):
+    config_dict = {
+        calculated_field: {
+            'input': input_field,
+            'algorithm': algorithm,
+            'type': aqi_type,
+        }
+    }
+    return config_dict
+
+class TestEPAAQICalculate(unittest.TestCase):
+    def test_get_series_data(self):
+        mock_logger = mock.Mock(spec=user.aqitype.Logger)
+
+        algorithm = 'EPAAQI'
+        aqi_type = 'pm2_5'
+        input_field = 'pm2_5'
+
+        calculated_field = random_string()
+
+        config_dict = setup_config(calculated_field, input_field, algorithm, aqi_type)
+        config = configobj.ConfigObj(config_dict)
+
+        with mock.patch.object(user.aqitype.EPAAQI, 'calculate', side_effect=mock_calculate_effect) as mock_calculate:
+            SUT = user.aqitype.AQIType(mock_logger, config)
+
+            unit = random_string()
+            unit_group = random_string()
+
+            with mock.patch('weewx.units.getStandardUnitType', return_value=[unit, unit_group]):
+                manager_config = configobj.ConfigObj(io.StringIO(MANAGER_DICT))
+                db_binder = weewx.manager.DBBinder(manager_config)
+                db_manager = db_binder.get_manager('aqi_binding')
+
+                timespan = weeutil.weeutil.TimeSpan(1740114000, 1740200400)
+
+                start_vec_t, stop_vec_t, _data_vec_t = SUT._get_series_epaaqi(calculated_field, timespan, db_manager, None, None)
+
+                mock_calculate.assert_called()
+                self.assertEqual(mock_calculate.call_count, 288)
+
+                i = 0
+                for call_arg in mock_calculate.call_args_list:
+                    self.assertEqual(call_arg[0][2], expected_pm_values[i])
+                    self.assertEqual(stop_vec_t[0][i], expected_timestamps[i])
+                    self.assertEqual(start_vec_t[0][i], expected_timestamps[i] - archive_interval_seconds)
+                    i += 1
+
+    def test_get_aggregate_avg_data(self):
+        mock_logger = mock.Mock(spec=user.aqitype.Logger)
+
+        algorithm = 'EPAAQI'
+        aqi_type = 'pm2_5'
+        input_field = 'pm2_5'
+
+        calculated_field = random_string()
+
+        config_dict = setup_config(calculated_field, input_field, algorithm, aqi_type)
+        config = configobj.ConfigObj(config_dict)
+
+        with mock.patch.object(user.aqitype.EPAAQI, 'calculate', side_effect=mock_calculate_effect    ) as mock_calculate:
+            SUT = user.aqitype.AQIType(mock_logger, config)
+
+            unit = random_string()
+            unit_group = random_string()
+
+            with mock.patch('weewx.units.getStandardUnitType', return_value=[unit, unit_group]):
+                manager_config = configobj.ConfigObj(io.StringIO(MANAGER_DICT))
+                db_binder = weewx.manager.DBBinder(manager_config)
+                db_manager = db_binder.get_manager('aqi_binding')
+
+                timespan = weeutil.weeutil.TimeSpan(1740114000, 1740200400)
+
+                _ret_value = SUT._get_aggregate_epaaqi(calculated_field, timespan, 'avg', db_manager)
+
+                self.assertEqual(mock_calculate.call_count, 288)
+
+                i = 0
+                for call_arg in mock_calculate.call_args_list:
+                    self.assertEqual(call_arg[0][2], expected_pm_values[i])
+                    i += 1
+
+    def test_get_aggregate_min_data(self):
+        mock_logger = mock.Mock(spec=user.aqitype.Logger)
+
+        algorithm = 'EPAAQI'
+        aqi_type = 'pm2_5'
+        input_field = 'pm2_5'
+
+        calculated_field = random_string()
+
+        config_dict = setup_config(calculated_field, input_field, algorithm, aqi_type)
+        config = configobj.ConfigObj(config_dict)
+
+        with mock.patch.object(user.aqitype.EPAAQI, 'calculate', return_value=random.randint(1, 100)) as mock_calculate:
+            SUT = user.aqitype.AQIType(mock_logger, config)
+
+            unit = random_string()
+            unit_group = random_string()
+
+            with mock.patch('weewx.units.getStandardUnitType', return_value=[unit, unit_group]):
+                manager_config = configobj.ConfigObj(io.StringIO(MANAGER_DICT))
+                db_binder = weewx.manager.DBBinder(manager_config)
+                db_manager = db_binder.get_manager('aqi_binding')
+
+                timespan = weeutil.weeutil.TimeSpan(1740114000, 1740200400)
+
+                _ret_value = SUT._get_aggregate_epaaqi(calculated_field, timespan, 'min', db_manager)
+
+                mock_calculate.assert_called_once_with(db_manager, None, min(expected_pm_values), aqi_type)
+
+if __name__ == '__main__':
+    #test_suite = unittest.TestSuite()
+    #test_suite.addTest(TestEPAAQICalculate('test_get_aggregate_min_data'))
+    #unittest.TextTestRunner().run(test_suite)
+
+    unittest.main(exit=False)
